fs/super.c: fix double prealloc_shrinker() in sget_fc()
I'm observing "KASAN: use-after-free Read in shrink_slab" on recent
linux-next in the code I've added:

https://syzkaller.appspot.com/bug?id=91767fc6346a4b9e0309a8cd7e2f356c434450b9

It seems to be not related to my patchset, since there is a problem with
double preallocation of shrinker.  We should use
register_shrinker_prepared() in sget_fc(), since shrinker is already
allocated in alloc_super().

Link: http://lkml.kernel.org/r/153131984019.24777.15284245961241666054.stgit@localhost.localdomain
Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Cc: David Howells <dhowells@redhat.com>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>