proc/kcore: don't grab lock for kclist_add()
Patch series "/proc/kcore improvements", v2.

This series makes a few improvements to /proc/kcore.  Patches 1 and 2 are
prep patches.  Patch 3 is a fix/cleanup.  Patch 4 is another prep patch. 
Patches 5 and 6 are optimizations to ->read().  Patch 7 adds vmcoreinfo to
/proc/kcore (apparently I'm not the only one who wants this, see
https://www.spinics.net/lists/arm-kernel/msg665103.html).

I tested that the crash utility still works with this applied, and readelf
is happy with it, as well.


This patch (of 7):

kclist_add() is only called at init time, so there's no point in grabbing
any locks.  We're also going to replace the rwlock with a rwsem, which we
don't want to try grabbing during early boot.

Link: http://lkml.kernel.org/r/ddbe991f42b8d49e50645878ad8c1a8168db2e82.1531440458.git.osandov@fb.com
Signed-off-by: Omar Sandoval <osandov@fb.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Alexey Dobriyan <adobriyan@gmail.com>
Cc: Eric Biederman <ebiederm@xmission.com>
Cc: James Morse <james.morse@arm.com>
Cc: Bhupesh Sharma <bhsharma@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>