staging: rtl8188eu/rtl8723bs: fix character encoding
drivers/staging/rtl8188eu/include/odm.h uses an incorrect encoding for the
'...' character in two comments, which makes it one of the few non-UTF-8
source files.

This removes the odd characters and uses the same ASCII representation
that we have in the regular rtlwifi driver.  The second instance of
drivers/staging/rtl8723bs/hal/odm.h is garbled in a different way, so I
change it to be the same as well even though it is already plain ASCII.

Link: http://lkml.kernel.org/r/20180724111600.4158975-4-arnd@arndb.de
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>