mm/memory_hotplug.c: clean up unregister_mem_sect_under_nodes()
With the assumption that the relationship between memory_block <-> node is
1:1, we can refactor this function a bit.

This assumption is being taken from register_mem_sect_under_node() code.

register_mem_sect_under_node() takes the mem_blk's nid, and compares it to
the pfn's nid we are checking.  If they match, we go ahead and link both
objects.  Once done, we just return.

So the relationship between memory_block <-> node seems to stand.

Currently, unregister_mem_sect_under_nodes() defines a nodemask_t which is
being checked in the loop to see if we have already unlinked a certain
node.  But since a memory_block can only belong to a single node, we can
drop the nodemask and the check within the loop.

If we find a match between the mem_block->nid and the nid of the pfn we
are checking, we unlink the objects and return, as unlinking the objects
once is enough.

Link: http://lkml.kernel.org/r/20180810152931.23004-4-osalvador@techadventures.net
Signed-off-by: Oscar Salvador <osalvador@suse.de>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Jonathan Cameron <jonathan.cameron@huawei.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: <yasu.isimatu@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>