mm/memory_hotplug.c: refactor unregister_mem_sect_under_nodes
unregister_mem_sect_under_nodes() tries to allocate a nodemask_t in order
to check within the loop which nodes have already been unlinked, so we do
not repeat the operation on them.

NODEMASK_ALLOC calls kmalloc() if NODES_SHIFT > 8, otherwise it just
declares a nodemask_t variable whithin the stack.

Since kmalloc() can fail, we actually check whether NODEMASK_ALLOC failed
or not, and we return -ENOMEM accordingly.  remove_memory_section() does
not check for the return value though.

The problem with this is that if we return -ENOMEM, it means that
unregister_mem_sect_under_nodes will not be able to remove the symlinks,
but since we do not check the return value, we go ahead and we call
unregister_memory(), which will remove all the mem_blks directories.

This will leave us with dangled symlinks.

The easiest way to overcome this is to fallback by calling
sysfs_remove_link() unconditionally in case NODEMASK_ALLOC failed.  This
means that we will call sysfs_remove_link on nodes that have been already
unlinked, but nothing wrong happens as sysfs_remove_link() backs off
somewhere down the chain in case the link has already been removed.

I think that this is better than

a) dangled symlinks
b) having to recovery from such error in remove_memory_section

Since from now on we will not need to care about return values, we can make
the function void.

As we have a safe fallback, one thing that could also be done is to add
__GFP_NORETRY in the flags when calling NODEMASK_ALLOC, so we do not retry.

Link: http://lkml.kernel.org/r/20180815144219.6014-4-osalvador@techadventures.net
Signed-off-by: Oscar Salvador <osalvador@suse.de>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Jonathan Cameron <jonathan.cameron@huawei.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: <yasu.isimatu@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>