tools/vm/page-types.c: include shared map counts
Add a new flag that will read kpagecount for each PFN and print out the
number of times the page is mapped along with the flags in the listing
view.

This information is useful in understanding and optimizing memory usage. 
Identifying pages which are not shared allows us to focus on adjusting the
memory layout or access patterns for the sole owning process.  Knowing the
number of processes that share a page tells us how many other times we
must make the same adjustments or how many processes to potentially
disable.

Truncated sample output:

voffset map-cnt offset  len     flags
561a3591e       1       15fe8   1       ___U_lA____Ma_b___________________________
561a3591f       1       2b103   1       ___U_lA____Ma_b___________________________
561a36ca4       1       2cc78   1       ___U_lA____Ma_b___________________________
7f588bb4e       14      2273c   1       __RU_lA____M______________________________

[akpm@linux-foundation.org: coding-style fixes]
[chansen3@cisco.com: add documentation, tweak whitespace]
  Link: http://lkml.kernel.org/r/20180705181204.5529-1-chansen3@cisco.com
Link: http://lkml.kernel.org/r/20180612153205.12879-1-chansen3@cisco.com
Signed-off-by: Christian Hansen <chansen3@cisco.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>